requirements.txt
# clinical-trial-repurposing-pipeline-core
# Python >=3.10 required

# ── Config & validation ────────────────────────────────────────────────────
pydantic>=2.0
PyYAML>=6.0

# ── CLI & display ──────────────────────────────────────────────────────────
typer>=0.9
rich>=13.0

# ── LLM / AI ──────────────────────────────────────────────────────────────
openai>=1.0              # OpenRouter uses OpenAI-compatible API
jinja2>=3.1              # last_mile template rendering

# ── HTTP ──────────────────────────────────────────────────────────────────
requests>=2.31
httpx>=0.27          # async-capable fallback

# ── Data manipulation ─────────────────────────────────────────────────────
pandas>=2.0
numpy>=1.26

# ── Biological databases ──────────────────────────────────────────────────
biopython>=1.83                  # PubMed / Entrez
chembl_webresource_client>=0.10  # ChEMBL REST wrapper

# ── Network biology ───────────────────────────────────────────────────────
networkx>=3.2                    # STRING PPI shortest path

# ── Statistics ────────────────────────────────────────────────────────────
scipy>=1.12

# ── Pathway enrichment ────────────────────────────────────────────────────
gseapy>=1.1

# ── LINCS L1000 (optional – only needed for local GEO .gctx files) ────────
# cmapPy>=4.0.1     # pip install cmapPy  (large; install separately)

# ── PDF dossier export (optional) ────────────────────────────────────────
# weasyprint>=60    # pip install weasyprint  (requires system libs)
# markdown>=3.6     # pip install markdown    (Markdown→HTML for weasyprint)
# reportlab>=4.0    # pip install reportlab   (pure-Python PDF alternative)

# ── Notebook / dev ────────────────────────────────────────────────────────
jupyter>=1.0       # optional; for exploratory notebooks
ipykernel>=6.0
